OS-1 Server/requirements.txt
pydantic==1.10.7
Pillow==8.4.0
pydub==0.25.1
pymongo==4.3.3
langchain==0.0.235
transformers==4.31.0
noisereduce==2.0.1
numpy==1.23.5
OpenCC==1.1.6
requests==2.26.0
gradio==3.35.2
gradio-client==0.2.9
typing-extensions==4.5.0
humanize==4.6.0
jax==0.4.10
torch==2.0.1
whisper==1.1.10
opencv-python==4.6.0.66
promptlayer==0.1.80
torchvision==0.15.2
PyYAML==5.4.1
paddlespeech==1.4.1
scipy==1.10.1
yacs==0.1.8
matplotlib==3.2.1
redis==4.5.4
pymilvus==2.2.8
fastapi==0.88.0
schedule==1.2.0
bcrypt==4.0.1
fschat[model_worker,webui]==0.2.29
sentence-transformers==2.2.2
openai
cryptography==41.0.4
aliyun-python-sdk-core-v3